automl/meta_trainer.py
"""
Meta-model training utilities for stacking ensemble methods.

Provides automated training of second-level models that learn to combine
predictions from multiple base models for improved accuracy.
"""

import asyncio
import logging
import time
from typing import Dict, List, Optional, Any, Tuple
import numpy as np
from pathlib import Path
import joblib

logger = logging.getLogger(__name__)


class MetaModelTrainer:
    """
    Automated meta-model trainer for stacking ensemble methods.
    
    Trains second-level models that learn optimal combinations of
    base model predictions for improved accuracy and calibration.
    """
    
    def __init__(self):
        self.available_meta_models = {
            "xgboost": self._train_xgboost,
            "logistic": self._train_logistic_regression,
            "random_forest": self._train_random_forest,
            "neural_network": self._train_neural_network,
            "lightgbm": self._train_lightgbm,
            "catboost": self._train_catboost
        }
        
        self.trained_models = {}
        self.training_history = []
    
    async def train_all_meta_models(
        self,
        base_predictions: List[List[float]],
        true_labels: List[int],
        validation_split: float = 0.2
    ) -> Dict[str, Any]:
        """
        Train all available meta-models and compare performance.
        
        Args:
            base_predictions: Predictions from all base models
            true_labels: Ground truth labels
            validation_split: Fraction of data for validation
            
        Returns:
            Dict: Training results and performance comparison
        """
        logger.info(f"Training {len(self.available_meta_models)} meta-models")
        
        # Split data for meta-model training
        split_idx = int(len(base_predictions) * (1 - validation_split))
        
        train_preds = base_predictions[:split_idx]
        train_labels = true_labels[:split_idx]
        val_preds = base_predictions[split_idx:]
        val_labels = true_labels[split_idx:]
        
        # Train all meta-models concurrently
        training_tasks = []
        for model_name, trainer_func in self.available_meta_models.items():
            task = trainer_func(train_preds, train_labels, val_preds, val_labels)
            training_tasks.append((model_name, task))
        
        # Collect training results
        training_results = {}
        
        for model_name, task in training_tasks:
            try:
                result = await task
                training_results[model_name] = result
                
                if result["success"]:
                    self.trained_models[model_name] = result["model"]
                    logger.info(f"✓ {model_name} meta-model trained: {result['score']:.4f}")
                else:
                    logger.error(f"✗ {model_name} meta-model training failed")
                
            except Exception as e:
                logger.error(f"Meta-model training error for {model_name}: {e}")
                training_results[model_name] = {
                    "success": False,
                    "error": str(e),
                    "score": 0.0
                }
        
        # Find best performing meta-model
        best_model = max(
            training_results.items(),
            key=lambda x: x[1].get("score", 0.0)
        )
        
        training_summary = {
            "timestamp": time.time(),
            "total_samples": len(base_predictions),
            "training_samples": len(train_preds),
            "validation_samples": len(val_preds),
            "models_trained": len([r for r in training_results.values() if r.get("success")]),
            "best_model": best_model[0],
            "best_score": best_model[1].get("score", 0.0),
            "all_results": training_results
        }
        
        self.training_history.append(training_summary)
        
        # Save best model
        await self._save_best_meta_model(best_model[0], self.trained_models[best_model[0]])
        
        return training_summary
    
    async def _train_xgboost(
        self,
        train_preds: List[List[float]],
        train_labels: List[int],
        val_preds: List[List[float]],
        val_labels: List[int]
    ) -> Dict[str, Any]:
        """Train XGBoost meta-model with hyperparameter optimization."""
        try:
            # TODO: Implement XGBoost training
            # import xgboost as xgb
            # from sklearn.model_selection import GridSearchCV
            # 
            # # Prepare data
            # X_train = np.array(train_preds)
            # y_train = np.array(train_labels)
            # X_val = np.array(val_preds)
            # y_val = np.array(val_labels)
            # 
            # # Define hyperparameter grid
            # param_grid = {
            #     'n_estimators': [50, 100, 200],
            #     'max_depth': [3, 6, 10],
            #     'learning_rate': [0.01, 0.1, 0.2],
            #     'subsample': [0.8, 0.9, 1.0]
            # }
            # 
            # # Grid search with cross-validation
            # xgb_model = xgb.XGBClassifier(random_state=42)
            # grid_search = GridSearchCV(
            #     xgb_model, param_grid, cv=5, scoring='roc_auc', n_jobs=-1
            # )
            # 
            # # Train model
            # grid_search.fit(X_train, y_train)
            # best_model = grid_search.best_estimator_
            # 
            # # Validate performance
            # val_predictions = best_model.predict_proba(X_val)[:, 1]
            # score = roc_auc_score(y_val, val_predictions)
            # 
            # return {
            #     "success": True,
            #     "model": best_model,
            #     "score": score,
            #     "best_params": grid_search.best_params_,
            #     "cv_score": grid_search.best_score_
            # }
            
            # Mock XGBoost training
            await asyncio.sleep(0.5)
            return {
                "success": True,
                "model": "MockXGBoostModel",
                "score": np.random.uniform(0.85, 0.95),
                "best_params": {
                    "n_estimators": 100,
                    "max_depth": 6,
                    "learning_rate": 0.1
                }
            }
            
        except Exception as e:
            logger.error(f"XGBoost training error: {e}")
            return {"success": False, "error": str(e), "score": 0.0}
    
    async def _train_logistic_regression(
        self,
        train_preds: List[List[float]],
        train_labels: List[int],
        val_preds: List[List[float]],
        val_labels: List[int]
    ) -> Dict[str, Any]:
        """Train logistic regression meta-model."""
        try:
            # TODO: Implement logistic regression training
            # from sklearn.linear_model import LogisticRegression
            # from sklearn.model_selection import GridSearchCV
            # 
            # X_train = np.array(train_preds)
            # y_train = np.array(train_labels)
            # X_val = np.array(val_preds)
            # y_val = np.array(val_labels)
            # 
            # # Hyperparameter grid
            # param_grid = {
            #     'C': [0.01, 0.1, 1.0, 10.0, 100.0],
            #     'penalty': ['l1', 'l2', 'elasticnet'],
            #     'solver': ['liblinear', 'saga']
            # }
            # 
            # # Grid search
            # lr_model = LogisticRegression(random_state=42, max_iter=1000)
            # grid_search = GridSearchCV(lr_model, param_grid, cv=5, scoring='roc_auc')
            # grid_search.fit(X_train, y_train)
            # 
            # # Validate
            # val_predictions = grid_search.best_estimator_.predict_proba(X_val)[:, 1]
            # score = roc_auc_score(y_val, val_predictions)
            
            await asyncio.sleep(0.3)
            return {
                "success": True,
                "model": "MockLogisticModel",
                "score": np.random.uniform(0.80, 0.90),
                "best_params": {"C": 1.0, "penalty": "l2"}
            }
            
        except Exception as e:
            return {"success": False, "error": str(e), "score": 0.0}
    
    async def _train_random_forest(
        self,
        train_preds: List[List[float]],
        train_labels: List[int],
        val_preds: List[List[float]],
        val_labels: List[int]
    ) -> Dict[str, Any]:
        """Train Random Forest meta-model."""
        await asyncio.sleep(0.4)
        return {
            "success": True,
            "model": "MockRandomForestModel",
            "score": np.random.uniform(0.82, 0.92),
            "best_params": {"n_estimators": 100, "max_depth": 10}
        }
    
    async def _train_neural_network(
        self,
        train_preds: List[List[float]],
        train_labels: List[int],
        val_preds: List[List[float]],
        val_labels: List[int]
    ) -> Dict[str, Any]:
        """Train neural network meta-model."""
        await asyncio.sleep(0.6)
        return {
            "success": True,
            "model": "MockNeuralNetModel",
            "score": np.random.uniform(0.83, 0.93),
            "best_params": {"hidden_layer_sizes": (64, 32), "learning_rate": 0.001}
        }
    
    async def _train_lightgbm(
        self,
        train_preds: List[List[float]],
        train_labels: List[int],
        val_preds: List[List[float]],
        val_labels: List[int]
    ) -> Dict[str, Any]:
        """Train LightGBM meta-model."""
        await asyncio.sleep(0.4)
        return {
            "success": True,
            "model": "MockLightGBMModel",
            "score": np.random.uniform(0.84, 0.94),
            "best_params": {"num_leaves": 31, "learning_rate": 0.1}
        }
    
    async def _train_catboost(
        self,
        train_preds: List[List[float]],
        train_labels: List[int],
        val_preds: List[List[float]],
        val_labels: List[int]
    ) -> Dict[str, Any]:
        """Train CatBoost meta-model."""
        await asyncio.sleep(0.5)
        return {
            "success": True,
            "model": "MockCatBoostModel",
            "score": np.random.uniform(0.85, 0.95),
            "best_params": {"iterations": 100, "depth": 6}
        }
    
    async def _save_best_meta_model(self, model_name: str, model: Any):
        """Save the best performing meta-model for production use."""
        try:
            models_dir = Path("models/meta")
            models_dir.mkdir(parents=True, exist_ok=True)
            
            model_path = models_dir / f"best_{model_name}_meta_model.pkl"
            
            # TODO: Save actual model
            # joblib.dump(model, model_path)
            
            # Mock save
            with open(model_path, 'w') as f:
                f.write(f"Mock saved model: {model_name}")
            
            logger.info(f"Best meta-model saved: {model_path}")
            
        except Exception as e:
            logger.error(f"Failed to save meta-model: {e}")


# Global trainer instance
meta_trainer = MetaModelTrainer()