requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Image processing
Pillow==10.1.0
python-magic==0.4.27

# Scientific computing
numpy==1.24.3

# AI/ML frameworks for real model inference
tensorflow==2.14.0
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.200

# Image processing and computer vision
opencv-python==4.8.1.78

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0

# TODO: Add when implementing advanced features
# scikit-learn==1.3.2  # For ensemble methods and meta-models
# xgboost==2.0.1       # For stacking meta-models
# matplotlib==3.8.2    # For evaluation visualizations
# seaborn==0.13.0      # For statistical plots
# shap==0.43.0         # For model explainability